refactor: 개별 발송 기능을 Dialog 컴포넌트 클래스로 분리
